# libs/expense-tracker-client/src/model/potTemplateUpdateDTO.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * expense_tracker
 * A REST Api that offers various endpoints for handling shared expenses.
 * OpenAPI spec version: 1.2.0
 */
import type { PotTemplateUpdateDTOCronExpression } from './potTemplateUpdateDTOCronExpression';
import type { PotTemplateUpdateDTODefaultCurrencyId } from './potTemplateUpdateDTODefaultCurrencyId';
import type { PotTemplateUpdateDTOName } from './potTemplateUpdateDTOName';

export interface PotTemplateUpdateDTO {
  cron_expression?: PotTemplateUpdateDTOCronExpression;
  default_currency_id?: PotTemplateUpdateDTODefaultCurrencyId;
  name?: PotTemplateUpdateDTOName;
}