# libs/expense-tracker-client/src/model/userListDTO.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * expense_tracker
 * A REST Api that offers various endpoints for handling shared expenses.
 * OpenAPI spec version: 1.2.0
 */

export interface UserListDTO {
  users: string[];
}