# openapi/expense_tracker_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "expense_tracker",
    "description": "A REST Api that offers various endpoints for handling shared expenses.",
    "contact": {
      "name": "Ciel1996"
    },
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "1.2.0"
  },
  "paths": {
...
    "/api/v1/template": {
      "get": {
        "tags": [
          "Templates"
        ],
        "summary": "Gets a list of all pot templates that the bearer is owning.",
        "operationId": "get_pot_templates",
        "responses": {
          "200": {
            "description": "The pot templates that the bearer owns.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PotTemplateDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Templates"
        ],
        "summary": "Creates a pot template from the given DTO for the bearer.",
        "operationId": "create_pot_template",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPotTemplateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The pot template has been created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PotTemplateDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/template/{template_id}": {
      "get": {
        "tags": [
          "Templates"
        ],
        "summary": "Gets a pot template by id that the bearer is owning.",
        "operationId": "get_pot_template_by_id",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "description": "Database id for the pot template.  ",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The pot template that the bearer owns and requested.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PotTemplateDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Templates"
        ],
        "summary": "Updates template name, template currency and cron expression if the user is the owner of the template.",
        "operationId": "update_template",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "description": "Database id for the pot template.  ",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PotTemplateUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "The template has been updated."
          },
          "403": {
            "description": "Indicates that the user is not authorized to update the given pot template."
          },
          "404": {
            "description": "Indicates that the desired pot template does not exists."
          },
          "409": {
            "description": "Indicates that the desired pot template can't be updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Templates"
        ],
        "summary": "Deletes the given template if the user is the owner of the template.",
        "operationId": "delete_pot_template",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "description": "Database id for the pot template.  ",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The pot template has been deleted."
          },
          "403": {
            "description": "Indicates that the user is not authorized to delete the given pot template."
          },
          "404": {
            "description": "Indicates that the desired pot template does not exists."
          },
          "409": {
            "description": "Indicates that the desired pot template can't be deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/template/{template_id}/users/add": {
      "put": {
        "tags": [
          "Templates"
        ],
        "summary": "Adds the given users to the given template if the calling user is the owner of the template.",
        "operationId": "add_users_to_template",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "description": "Database id for the pot template.  ",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserListDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "The users has been added to the pot template."
          },
          "403": {
            "description": "Indicates that the users is not authorized to add users to the given pot template."
          },
          "404": {
            "description": "Indicates that the desired users or pot template does not exists."
          },
          "409": {
            "description": "Indicates that the desired users can't be added to the pot template."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/template/{template_id}/users/remove": {
      "put": {
        "tags": [
          "Templates"
        ],
        "summary": "Removes the given users from the given template if the calling user is the owner of the template.",
        "operationId": "remove_users_from_template",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "description": "Database id for the pot template.  ",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserListDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "The users have been removed from the pot template."
          },
          "403": {
            "description": "Indicates that the user is not authorized to remove users from the given pot template."
          },
          "404": {
            "description": "Indicates that the desired users or pot template does not exists."
          },
          "409": {
            "description": "Indicates that the desired users can't be removed from the pot template."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurrencyDTO": {
        "type": "object",
        "description": "DTO representing a currency.",
        "required": [
          "id",
          "name",
          "symbol"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          }
        }
      },
...
      "NewPotTemplateDTO": {
        "type": "object",
        "required": [
          "name",
          "default_currency_id",
          "user_ids",
          "cron_expression"
        ],
        "properties": {
          "cron_expression": {
            "type": "string",
            "description": "A cron expression that defines when the pot should be automatically created.\nThe example shows how a cron expression must look that expresses \"At 12:00 AM, on day 1 of the month\"\nI recommend using tools like https://crontab.cronhub.io to generate and validate a cron expression.\nExpense-Tracker uses the local timezone to make it easier for the user to think in their local timezone.",
            "example": "0 0 0 1 * *"
          },
          "default_currency_id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Name for the pot template. Supports placeholders for the current month {month} and year {year}.",
            "example": "My Pot {month}.{year}",
            "maxLength": 24
          },
          "user_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "A list of user ids that should be automatically added as members of this pot.\nThe owner does not need to be part of this list, as they are automatically added\nby the service."
          }
        }
      },
...
      "PotTemplateDTO": {
        "type": "object",
        "required": [
          "id",
          "owner",
          "name",
          "default_currency",
          "cron_expression",
          "users"
        ],
        "properties": {
          "cron_expression": {
            "type": "string",
            "description": "The example shows how a cron expression must look that expresses \"At 12:00 AM, on day 1 of the month\"\nI recommend using tools like https://crontab.cronhub.io to generate and validate a cron expression.\nExpense-Tracker uses the local timezone to make it easier for the user to think in their local timezone.",
            "example": "0 0 0 1 * *"
          },
          "default_currency": {
            "$ref": "#/components/schemas/CurrencyDTO"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDTO"
            }
          }
        }
      },
      "PotTemplateUpdateDTO": {
        "type": "object",
        "properties": {
          "cron_expression": {
            "type": [
              "string",
              "null"
            ]
          },
          "default_currency_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UserDTO": {
        "type": "object",
        "description": "The DTO representing a user from DB.",
        "required": [
          "uuid",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "uuid": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "UserListDTO": {
        "type": "object",
        "required": [
          "users"
        ],
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}